Separate webpack chunks for different environments (#238)
Currently if the feature has multiple environment files with the same target,
e.g. `feature.main.env.tsx` and `feature.preview.env.tsx` with the `web` target,
both are included in the same Webpack chunk. Meaning that both environments
unnecessarily load all the code designated for another environment.

Even though this unused code is not evaluated, it increases parse time and
memory consumption.

This PR creates one chunk for the `.feature` file, and one chunk for each
`.env` file.

In the project I'm working on this separation increases the total uncompressed
amount of code in the main environment from 19.5MB to 19.9MB, and decreases the
amount of code in the preview environment from 19.0MB to 1.8MB

There's still a lot of remaining overhead, for example many of `.feature`
chunks include the source code of all the features they depend on. That's
roughly 1MB of duplicate code.