feat(core): preserve original error message in cross-environment API calls
Sometimes we want to present the error message of a remote API call to the user
```
try {
    await service.doThing();
} catch (error) {
    alert(error.message);
}
```

But currently we prefix each error with 'Remote call failed with error:', and
also include the entire stack trace in the error message, which doesn't look
good in the UI.

Additionally, we discard custom error properties, such as error code.

This PR changes error handing to preserve the error name, message, and custom
properties. If the error is uncaught, it's still useful to know from which
environment it originated. This extra piece of information is added to the
`error.stack`. Chrome logs it in full to the console. Unfortunately, Firefox
ignores `error.stack` and shows its own stack trace that's not derived from the
error object.