feat(scripts): allow to resolve local paths on app.run
currently, because how we generate the manifest, we assume that the engine application will run from node_modules context.
Because of that it references own files using their "scoped name" to the package.json's name.
If the entrypoint of the application is not node_modules, it will fail at evaluation time when running the feature.

allowing to provide resolveLocalNodePaths flag in **application.run**, which will try to resolve own engine files (.feature., .env. etc..) from  the current base path + provided in build libOutDir (either from engine config or as a flag).

+ renaming featureOutDir to libOutDir and deprecating featureOutDir, since it doesn't properly describe this flag