feat: allow communication between node environment Currently multiple node environments can run, but they can't communicate.
Adding the ability to initiate communication between node environment using either the socket server initializer, or using the new `localNodeEnvironmentInitializer` being exported from the  `@wixc3/engine-core-node` package.

To support the localNodeEnvironmentInitializer, the node environment manager initiates a communication instance, and behaves as a middleman for communication between the node environment.

To support the socketInitializer, the way we load the environments changed a bit.
Now, we first understand the topology of all environments, so we first initiate all http servers, and only after, we launch the environments.
This ensures that all environment will be loaded with the complete topology.