test-kit(feat): add better tracing capabilities
Allow providing TRACING_PATH environment variable, when where it's passed, it will automatically save a trace for all running tests.
Also, as he value it is possible to provide where the traces should be saved (default is `process.cwd()`).

Allow providing specifying tracing in the suite level.

both these approaches will use the name of the test as the zip name