feat(test-kit): add allowedErrors for withFeature (#993)
naive mechanism to allow certain error messages to not fail the tests. others will still fail as expected.

this is *not* as flexible as the previous handler api, and that's actually on purpose. we'd like to lean towards the strict side as much as possible to avoid "real" unhandled logic errors (as opposed to ones our applications want to passthrough on purpose).